+++ tasks/forever-task - orig.js
@@ -3,8 +3,6 @@ var forever     = require('forever'),
     logDir      = path.join(process.cwd(), '/forever'),
     logFile     = path.join(logDir, '/out.log'),
     errFile     = path.join(logDir, '/err.log'),
-	foreverOpts = [],
-	paramsMatch = {},
     commandName = 'node',
     commandMap  = {
       start:      startForeverWithIndex,
@@ -45,12 +43,10 @@ function prettyPrint( id, object ) {
 /**
  * Locates running process previously started by forever based on index file, and notifies callback. Will notify of undefined if not found, other wise the unformatted process object.
  * @param  {String}   index    Index filename.
- @param {Object} params Additional parameters for identifying the process (in case have multiple with the same index / command)
-	@param {Array} optionsMatch Array of options to search through to try to match; each is a string
  * @param  {Function} callback Delegate method to invoke with either the found process object or undefined if not found.
  */
-function findProcessWithIndex( index, params, callback ) {
-  var i, jj, kk, process, uid =false;
+function findProcessWithIndex( index, callback ) {
+  var i, process;
   try {
     forever.list(false, function(context, list) {
       i = list ? list.length : 0;
@@ -58,29 +54,12 @@ function findProcessWithIndex( index, params, callback ) {
         process = list[i];
         if( process.hasOwnProperty('file') &&
           process.file === index ) {
-			// if(params.optionsMatch !==undefined) {
-				// if(process.hasOwnProperty('options')) {
-					// for(jj =0; jj<process.options.length; jj++) {
-						// for(kk =0; kk<params.optionsMatch.length; kk++) {
-							// if(process.options[jj].indexOf(params.optionsMatch[kk]) >-1) {
-								// uid =process.uid;
-								// break;
-							// }
-						// }
-					// }
-				// }
-			// }
-			// else {	//if no options to check, match on file is good enough
-				// uid =process.uid;
-				// break;
-			// }
-			uid =process.uid;
-			break;
+          break;
         }
         process = undefined;
       }
 
-      callback.call(null, process, uid);
+      callback.call(null, process);
     });
   }
   catch( e ) {
@@ -96,7 +75,7 @@ function startForeverWithIndex( index ) {
   log( 'Attempting to start ' + index + ' as daemon.');
 
   done = this.async();
-  findProcessWithIndex( index, paramsMatch, function(process, uid) {
+  findProcessWithIndex( index, function(process) {
     // if found, be on our way without failing.
     if( typeof process !== 'undefined' ) {
       warn( index + ' is already running.');
@@ -111,8 +90,7 @@ function startForeverWithIndex( index ) {
         outFile: logFile,
         command: commandName,
         append: true,
-        max: 3,
-		options: foreverOpts
+        max: 3
       });
       log( 'Logs can be found at ' + logDir + '.' );
       done();
@@ -127,17 +105,16 @@ function stopOnProcess(index) {
   log( 'Attempting to stop ' + index + '...' );
 
   done = this.async();
-  findProcessWithIndex( index, paramsMatch, function(process, uid) {
-    if( typeof process !== 'undefined' && uid ) {
+  findProcessWithIndex( index, function(process) {
+    if( typeof process !== 'undefined' ) {
       log( forever.format(true,[process]) );
 
-		// forever.stop( index )
-		forever.stop( uid )		//more specific
+      forever.stop( index )
         .on('stop', function() {
           done();
         })
         .on('error', function(message) {
-          error( 'Error stopping uid: ' + uid + 'index: ' + index + '. [REASON] :: ' + message );
+          error( 'Error stopping ' + index + '. [REASON] :: ' + message );
           done(false);
         });
     }
@@ -162,14 +139,13 @@ function restartOnProcess( index ) {
   }(this, index));
 
   done = this.async();
-  findProcessWithIndex( index, paramsMatch, function(process, uid) {
-    if(typeof process !== 'undefined' && uid) {
+  findProcessWithIndex( index, function(process) {
+    if(typeof process !== 'undefined') {
       log(forever.format(true,[process]));
 
-		// forever.restart( index)
-		forever.restart(uid)		//more specific
+      forever.restart( index)
         .on('error', function(message) {
-          error('Error restarting uid: '+uid+' index: ' + index + '. [REASON] :: ' + message);
+          error('Error restarting ' + index + '. [REASON] :: ' + message);
           done(false);
         });
       done();
@@ -200,13 +176,6 @@ module.exports = function(grunt) {
         logFile = path.join(logDir, this.options().logFile || 'out.log');
         errFile = path.join(logDir, this.options().errFile || 'err.log');
       }
-
-		foreverOpts =this.options().options || [];
-		paramsMatch =this.options().paramsMatch || {};
-		//default to match passed in options (if they exist)
-		if(paramsMatch.optionsMatch ===undefined) {
-			paramsMatch.optionsMatch =options;
-		}
 
       try {
         if(commandMap.hasOwnProperty(operation)) {